samples/a2a/business_agent/pyproject.toml
[project]
name = "business_agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic[email]>=2.12.3",
    "google-adk[a2a]>=1.22.0",
    "uvicorn>=0.35.0",
    "httpx>=0.28.1",
    "ucp-sdk",
]

[project.scripts]
business_agent = "business_agent.main:run"

[dependency-groups]
dev = [
    "ruff>=0.14.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
ucp-sdk = { path = "../../../sdk/python/", editable = true }
